- EMAR-227: Patient API updates for additional fields: image, numbers Emar.Api\Helpers\MediaTypes.cs Emar.Core\Helpers\Utilities.cs Emar.Core\Patients\Model\Constants.cs Emar.Core\Patients\Model\PatientDto.cs Emar.Core\Patients\Model\Mappings\PatientMapper.cs Emar.Core\Patients\Repository\IPatientRepository.cs Emar.Core\Patients\Repository\PatientRepository.cs Emar.Core\Patients\Service\IPatientService.cs Emar.Core\Patients\Service\PatientService.cs Emar.Data\Entities\ExternalId.cs Emar.Data\Entities\Patient.cs Emar.Data\Entities\Site.cs
- No ticket: Switch SiteId from int to long
Emar.Core\Medications\Model\MedicationRouteDto.cs
Emar.Core\Medications\Model\MedicationUnitDto.cs
Emar.Core\Options\Model\OptionDto.cs
Emar.Core\Options\Repository\IOptionRepository.cs
Emar.Core\Options\Repository\OptionRepository.cs
Emar.Core\Options\Service\IOptionService.cs
Emar.Core\Options\Service\OptionService.cs
Emar.Core\Orders\Model\UserQuickListItemDto.cs
Emar.Core\Orders\Repository\IOrderRepository.cs
Emar.Core\Orders\Repository\OrderRepository.cs
Emar.Core\Orders\Service\IOrderService.cs
Emar.Core\Orders\Service\OrderService.cs
Emar.Core\Orders\Service\OrderService_MockMethods.cs
Emar.Core\Patients\Model\PatientDto.cs
Emar.Core\ResourceParameters\PatientsResourceParameters.cs
Emar.Core\Sites\Model\SiteDto.cs
Emar.Data\Entities\Action.cs
Emar.Data\Entities\GroupListItem.cs
Emar.Data\Entities\MedicationRoute.cs
Emar.Data\Entities\MedicationUnit.cs
Emar.Data\Entities\Patient.cs
Emar.Data\Entities\Site.cs
Emar.Data\Entities\SiteOption.cs
Emar.Data\Entities\User.cs
Emar.Data\Entities\UserQuickListItem.cs

- No ticket: Cleanup
Emar.Data\EmarContext.cs

git-tfs-id: [http://ros-tfs-px01.picis.com:8080/tfs/emar]$/;C335

+++ WebAPI/API/C_Development/C_DevTrunk/Emar.Core/Orders/Model/UserQuickListItemDto.cs
@@ -13,7 +13,7 @@ namespace Emar.Core.Orders.Model
         /// <summary>
         /// User Quick Lists are site-specific
         /// </summary>
-        public int SiteId { get; set; }
+        public long SiteId { get; set; }
 
         public IEnumerable<HateOasLinkDto> Links;
     }
