+++ WebAPI/API/C_Development/C_DevTrunk/Emar.Data/Entities/Action.cs
@@ -24,7 +24,7 @@ namespace Emar.Data.Entities
         [StringLength(100)]
         public string Description { get; set; }
         [Column("site_id")]
-        public int SiteId { get; set; }
+        public long SiteId { get; set; }
         [Column("is_active")]
         public bool IsActive { get; set; }
