+++ WebAPI/API/C_Development/C_DevTrunk/Emar.Data/Entities/UserQuickListItem.cs
@@ -18,7 +18,7 @@ namespace Emar.Data.Entities
         public int Id { get; set; }
 
         [Column("site_id", TypeName = "int"), Required]
-        public int SiteId { get; set; }
+        public long SiteId { get; set; }
 
         [Column("user_id", TypeName = "int"), Required]
         public int UserId { get; set; }
